adding base_docstring and __internal__
