+++ keras_cv/models/darknet.py
@@ -20,170 +20,47 @@ Reference:
 
 import tensorflow as tf
 from tensorflow import keras
-from tensorflow.keras import backend
 from tensorflow.keras import layers
 
+from keras_cv.models.__internal__.darknet_utils import DarknetConvBlock
+from keras_cv.models.__internal__.darknet_utils import ResidualBlocks
+from keras_cv.models.__internal__.darknet_utils import SPPBottleneck
 
-def DarknetConvBlock(
-    filters,
-    kernel_size,
-    strides,
-    groups=1,
-    use_bias=False,
-    activation="silu",
-    name=None,
-):
-    """The basic conv block used in Darknet. Applies Conv2D followed by a BatchNorm.
+BASE_DOCSTRING = """Instantiates the {name} architecture.
 
-    Args:
-        filters: Integer, the dimensionality of the output space (i.e. the number of
-            output filters in the convolution).
-        kernel_size: An integer or tuple/list of 2 integers, specifying the height
-            and width of the 2D convolution window. Can be a single integer to specify
-            the same value both dimensions.
-        strides: An integer or tuple/list of 2 integers, specifying the strides of
-            the convolution along the height and width. Can be a single integer to
-            the same value both dimensions.
-        groups: A positive integer specifying the number of groups in which the
-            input is split along the channel axis. Each group is convolved separately
-            with `filters / groups` filters. The output is the concatenation of all
-            the `groups` results along the channel axis. Input channels and `filters`
-            must both be divisible by `groups`.
-        use_bias: Boolean, whether the layer uses a bias vector.
-        activation: the activation applied after the BatchNorm layer. One of "silu",
-            "relu" or "lrelu". Defaults to "silu".
-        name: the prefix for the layer names used in the block.
+    Although the {name} architecture is commonly used for detection tasks, it is
+    possible to extract the intermediate dark2 to dark5 layers from the model for
+    creating a feature pyramid Network.
 
-    Returns:
-        a function that takes an input Tensor representing a DarknetConvBlock.
-    """
-
-    if name is None:
-        name = f"darknet_block_{backend.get_uid('darknet_block')}"
-
-    def apply(x):
-        x = layers.Conv2D(
-            filters,
-            kernel_size,
-            strides,
-            padding="same",
-            groups=groups,
-            use_bias=use_bias,
-            name=name,
-        )(x)
-        x = layers.BatchNormalization(name=f"{name}_bn")(x)
-
-        if activation == "silu":
-            x = keras.activations.swish(x)
-        elif activation == "relu":
-            x = layers.ReLU(name=f"{name}_relu")(x)
-        elif activation == "lrelu":
-            x = layers.LeakyReLU(0.1, name=f"{name}_lrelu")(x)
-
-        return x
-
-    return apply
-
-
-def ResidualBlocks(filters, num_blocks, name=None):
-    """A residual block used in DarkNet models, repeated `num_blocks` times.
+    Reference:
+        - [YoloV3 Paper](https://arxiv.org/abs/1804.02767)
+        - [YoloV3 implementation](https://github.com/ultralytics/yolov3)
+    For transfer learning use cases, make sure to read the [guide to transfer learning
+    & fine-tuning](https://keras.io/guides/transfer_learning/).
 
     Args:
-        filters: Integer, the dimensionality of the output spaces (i.e. the number of
-            output filters in used the blocks).
-        num_blocks: number of times the residual connections are repeated
-        name: the prefix for the layer names used in the block.
-
+        include_rescaling: whether or not to Rescale the inputs.If set to True,
+            inputs will be passed through a `Rescaling(1/255.0)` layer.
+        include_top: whether to include the fully-connected layer at the top of
+            the network.  If provided, `classes` must be provided.
+        classes: optional number of classes to classify images into, only to be
+            specified if `include_top` is True, and if no `weights` argument is
+            specified.
+        weights: one of `None` (random initialization), or a pretrained weight
+            file path.
+        input_shape: optional shape tuple, defaults to (None, None, 3).
+        pooling: optional pooling mode for feature extraction when `include_top`
+            is `False`.
+            - `None` means that the output of the model will be the 4D tensor output
+                of the last convolutional block.
+            - `avg` means that global average pooling will be applied to the
+                output of the last convolutional block, and thus the output of the
+                model will be a 2D tensor.
+            - `max` means that global max pooling will be applied.
+        name: (Optional) name to pass to the model.  Defaults to "{name}".
     Returns:
-        a function that takes an input Tensor representing a ResidualBlock.
-    """
-
-    if name is None:
-        name = f"residual_block_{backend.get_uid('residual_block')}"
-
-    def apply(x):
-        x = DarknetConvBlock(
-            filters, kernel_size=3, strides=2, activation="lrelu", name=f"{name}_conv1"
-        )(x)
-
-        for i in range(1, num_blocks + 1):
-            residual = x
-
-            x = DarknetConvBlock(
-                filters // 2,
-                kernel_size=1,
-                strides=1,
-                activation="lrelu",
-                name=f"{name}_conv{2*i}",
-            )(x)
-            x = DarknetConvBlock(
-                filters,
-                kernel_size=3,
-                strides=1,
-                activation="lrelu",
-                name=f"{name}_conv{2*i + 1}",
-            )(x)
-
-            if i == num_blocks:
-                x = layers.Add(name=f"{name}_out")([residual, x])
-            else:
-                x = layers.Add(name=f"{name}_add_{i}")([residual, x])
-
-        return x
-
-    return apply
-
-
-def SPPBottleneck(filters, kernel_sizes=(5, 9, 13), activation="silu", name=None):
-    """Spatial pyramid pooling layer used in YOLOv3-SPP
-
-    Args:
-        filters: Integer, the dimensionality of the output spaces (i.e. the number of
-            output filters in used the blocks).
-        kernel_sizes: A list or tuple representing all the pool sizes used for the
-            pooling layers. Defaults to (5, 9, 13).
-        activation: Activation for the conv layers. Defaults to "silu".
-        name: the prefix for the layer names used in the block.
-
-    Returns:
-        a function that takes an input Tensor representing an SPPBottleneck.
-    """
-
-    if name is None:
-        name = f"spp_{backend.get_uid('spp')}"
-
-    def apply(x):
-        x = DarknetConvBlock(
-            filters,
-            kernel_size=1,
-            strides=1,
-            activation=activation,
-            name=f"{name}_conv1",
-        )(x)
-        x = [x]
-
-        for kernel_size in kernel_sizes:
-            x.append(
-                layers.MaxPooling2D(
-                    kernel_size,
-                    strides=1,
-                    padding="same",
-                    name=f"{name}_maxpool_{kernel_size}",
-                )(x[0])
-            )
-
-        x = layers.Concatenate(name=f"{name}_concat")(x)
-        x = DarknetConvBlock(
-            filters,
-            kernel_size=1,
-            strides=1,
-            activation=activation,
-            name=f"{name}_conv2",
-        )(x)
-
-        return x
-
-    return apply
+        A `keras.Model` instance.
+"""
 
 
 def DarkNet(
@@ -214,9 +91,8 @@ def DarkNet(
         blocks: numbers of building blocks from the layer dark2 to layer dark5.
         include_rescaling: whether or not to Rescale the inputs.If set to True,
             inputs will be passed through a `Rescaling(1/255.0)` layer.
-        include_top: whether to include the fully-connected
-            layer at the top of the network.  If provided, `classes` must be
-            provided.
+        include_top: whether to include the fully-connected layer at the top of
+            the network.  If provided, `classes` must be provided.
         classes: optional number of classes to classify imagesinto, only to be
             specified if `include_top` is True, and if no `weights` argument is
             specified.
@@ -321,7 +197,7 @@ def DarkNet21(
     weights=None,
     input_shape=(None, None, 3),
     pooling=None,
-    name="darknet21",
+    name="DarkNet21",
     **kwargs,
 ):
     return DarkNet(
@@ -344,7 +220,7 @@ def DarkNet53(
     weights=None,
     input_shape=(None, None, 3),
     pooling=None,
-    name="darknet53",
+    name="DarkNet53",
     **kwargs,
 ):
     return DarkNet(
@@ -360,5 +236,5 @@ def DarkNet53(
     )
 
 
-setattr(DarkNet21, "__doc__", DarkNet.__doc__)
-setattr(DarkNet53, "__doc__", DarkNet.__doc__)
+setattr(DarkNet21, "__doc__", BASE_DOCSTRING.format(name="DarkNet21"))
+setattr(DarkNet53, "__doc__", BASE_DOCSTRING.format(name="DarkNet21"))
